FEATURES ADDITION: SRI scene, notif's, minor fixes
Problem
The game has to have SRI to let user to upgrade
factories and browse through the unlocked facts.
The game should notify user about unlocking a
new planet or a new fact.

Solution
SRI scene has been added. Upgrades working perfectly.
The cost and the power of an upgrade can be adjusted
in the GameDataManager object.
Notifications have been added to the result scene.

Note
Some minor bugs have been fixed.

+++ Spaceport/ControlTowerManager.cs
@@ -135,9 +135,9 @@ public class ControlTowerManager : MonoBehaviour {
         GameDataManager.instance.SaveToDataFile("/PlanetData/planet_" + localData.index.ToString() + ".json", localData);
     }
 
-    private float CalculateFuelConsumption()
+    private int CalculateFuelConsumption()
     {
-        return GameDataManager.instance.planetData[selectedPlanet].fuelToReach * (1f - GameDataManager.instance.GetEngineBonus());
+        return (int)(GameDataManager.instance.planetData[selectedPlanet].fuelToReach * (1f - GameDataManager.instance.GetEngineBonus()));
     }
 
 }
